Fix redirect/authentication loop when SAML transaction fails
The /auth/failure endpoint was not reachable due to the order handlers
are defined. Rather than rearranging them, for now I've added a `pass`
in the catch-all handler for this.

It might be a better solution long-term to just move the catch all last
to avoid the need for any use of `pass`.

Resolves #1