Use a random cookie secret at runtime if not provided
Generating a random cookie secret is more secure than having a static
string in the code, and it makes it easier to automate deployment of
okta-auth-proxy. There's no trade-off for doing this as we'll
automatically be reauthenticated with the SAML provider if the auth
proxy is restarted and our current cookie becomes unreadable.